% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilSpecifyLesionWeightsDistr.R
\name{Convert2lesDistr}
\alias{Convert2lesDistr}
\title{Convert the 1D lesion distribution array to the standard 2D form}
\usage{
Convert2lesDistr(specified1DLesDistr)
}
\arguments{
\item{specified1DLesDistr}{The specified lesion distribution array. For example, 
c(0.1, 0.2, 0, 0.7) specifies 10 percent of diseased cases have one lesion,
20 percent have two lesions, none have 3 lesions and 70 percent have 
four lesions. See example below.}
}
\value{
lesDistr The lesion distribution array in standard 2D form
}
\description{
{#' Specify lesion weight distribution array.}
}
