% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset02}
\alias{dataset02}
\title{Van Dyke ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:2, 1:5, 1:114, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:2, 1:5, 1:45, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:45], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:45, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:45, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:2] "0" "1", treatment labels}
\item{\code{readerID}}{, chr [1:5] "0" "1" "2" ..., reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:69] 1 2 3 4 5 6 7 8 9 10 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:45] 70 71 72 73 74 75 76 77 78 79 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:2, 1:5, 1:114, 1:2] 1 1 1 1 1  ..., truth table structure}
\item{\code{datasetName}}{"VanDyke"}
}
}
\usage{
dataset02
}
\description{
This is referred to in the book as the "VD" dataset. It consists of 114 cases,
   45 of which are diseased, interpreted in two treatments ("0" = single spin echo 
   MRI, "1" = cine-MRI) by five radiologists using the ROC 
   paradigm. Each diseased cases had an aortic dissection; the ROC paradigm 
   generates one rating per case.  Often referred to in the ROC literature as the 
   Van Dyke dataset, which, along with the Franken dataset, has been widely 
   used to illustrate advances in ROC methodology. The example below displays 
   the ROC plot for the first treatment and first reader.
}
\examples{
str(dataset02)
PlotEmpiricalOperatingCharacteristics(dataset = dataset02, opChType = "ROC")$Plot



}
\references{
Van Dyke CW, et al. Cine MRI in the diagnosis of thoracic 
aortic dissection. 79th RSNA Meetings. 1993.
}
\keyword{datasets}
