% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset03}
\alias{dataset03}
\title{Franken ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:2, 1:4, 1:100, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:2, 1:4, 1:67, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:67], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:67, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:67, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:2] "TREAT1" "TREAT2", the treatment labels}
\item{\code{readerID}}{, chr [1:4] "READER_1" "READER_2" "READER_3" "READER_4", the reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:33] 1 2 3 4 5 6 7 8 9 10 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:67] 34 35 36 37 38 39 40 41 42 43 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:2, 1:4, 1:100, 1:2] 1 1 1 1 1 1 1   ..., truth table structure}
}
}
\usage{
dataset03
}
\description{
This is referred to in the book as the "FR" dataset. It consists of 100 cases, 
67 of which are diseased, interpreted in two treatments,
"0" = conventional film radiographs, "1" = digitized  images viewed on monitors, by four 
radiologists using the ROC paradigm. Often referred to in the ROC literature as the 
Franken-dataset, which, along the the Van Dyke dataset, has been widely used to illustrate
advances in ROC methodology.
}
\examples{
str(dataset03)
PlotEmpiricalOperatingCharacteristics(dataset = dataset03, opChType = "ROC")$Plot


}
\references{
Franken EA, et al. Evaluation of a Digital Workstation for Interpreting 
Neonatal Examinations: A Receiver Operating Characteristic Study. Investigative Radiology. 
1992;27(9):732-737.
}
\keyword{datasets}
