% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfLroc2Roc.R
\name{DfLroc2Roc}
\alias{DfLroc2Roc}
\title{Convert an LROC dataset to a ROC dataset}
\usage{
DfLroc2Roc (dataset)
}
\arguments{
\item{dataset}{The \strong{LROC} dataset to be converted.}
}
\value{
An ROC dataset
}
\description{
Converts an LROC dataset to a ROC dataset
}
\details{
The conversion is effected by taking the maximum rating on each diseased case, 
which could be a TPCl or a TPIl, whichever has the higher rating.
}
\examples{
dataset <- DfReadLrocDataFile()
str(dataset)
rocDataSet <- DfLroc2Roc(dataset)
str(rocDataSet)


}
