% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfReadSP_A.R
\name{DfReadSP_A}
\alias{DfReadSP_A}
\title{Read a split plot data file}
\usage{
DfReadSP_A(fileName)
}
\arguments{
\item{fileName}{String specifying the name of the file.}
}
\value{
A dataset with the structure specified in
  \code{\link{RJafroc-package}}.
}
\description{
Read disk file and create a dataset object
}
\examples{
fileName <- system.file("extdata", "toyFiles/ROC/rocSpA.xlsx",
package = "RJafroc", mustWork = TRUE)
ds <- DfReadSP_A(fileName)


}
