% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfWriteExcelDataFile.R
\name{DfWriteExcelDataFile}
\alias{DfWriteExcelDataFile}
\title{Save dataset object as a JAFROC format Excel file}
\usage{
DfWriteExcelDataFile(dataset, fileName)
}
\arguments{
\item{dataset}{{The dataset object, see \code{\link{RJafroc-package}}}.}

\item{fileName}{{The file name to save to; the extension of the data file must be .xlsx}}
}
\description{
Save a dataset object as a JAFROC format Excel file
}
\examples{
\donttest{
##DfWriteExcelDataFile(dataset = dataset05, fileName = "rocData2.xlsx")
}

}
