% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FROC2HrROC.R
\name{FROC2HrROC}
\alias{FROC2HrROC}
\title{Convert FROC dataset}
\usage{
FROC2HrROC(dataset)
}
\arguments{
\item{dataset}{The dataset to be converted, see \link{RJafroc-package}.}
}
\value{
An ROC dataset, where each case is represented by the highest rating of the case in FROC dataset.
}
\description{
Convert an FROC dataset to a highest rating inferred ROC dataset
}
\examples{
FROC2HrROC(dataset = frocData)
}

