% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2Functions.R
\name{funs}
\alias{funs}
\alias{sum1}
\alias{prod1}
\title{Two functions of x and y}
\usage{
funs(x, y)

sum1(x, y)

prod1(x, y)
}
\arguments{
\item{x}{=X}

\item{y}{=Y}
}
\value{
\code{sum1(x,y)} returns x+y

\code{prod1(x,y)} returns x*y
}
\description{
Two functions of x and y
}
\note{
\code{funs} is a generic name for the functions documented.
\cr
If called, \code{funs} returns its own arguments.
}
\examples{
sum1(3,4)
prod1(3,4)
}
