% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfBinDataset.R
\name{isBinnedDataset}
\alias{isBinnedDataset}
\title{Determine if a dataset is binned}
\usage{
isBinnedDataset(dataset, maxUniqeRatings = 6)
}
\arguments{
\item{dataset}{The dataset}

\item{maxUniqeRatings}{The max number of unique ratings in order to be classified as binned, default is 6}
}
\value{
logical I x J array, TRUE if the corresponding treatment-reader dataset is binned, 
   i.e., has at most 6 unique levels, FALSE otherwise
}
\description{
Determine if a dataset is binned
}
\examples{
isBinnedDataset(dataset01)

}
