% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDataset.R
\name{isValidDataset}
\alias{isValidDataset}
\title{Check the validity of a dataset}
\usage{
isValidDataset(dataset)
}
\arguments{
\item{dataset}{The dataseet object to be checked.}
}
\value{
\code{TRUE} if dataset is valid, \code{FALSE} otherwise.
}
\description{
Checks the validity of the dataset.
}
