% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBinormalFit.R
\name{PlotBinormalFit}
\alias{PlotBinormalFit}
\title{Plot binormal fit}
\usage{
PlotBinormalFit(a, b)
}
\arguments{
\item{a}{vector: the mean(s) of the diseased distribution(s).}

\item{b}{vector: the standard deviations(s) of the diseased distribution(s).}
}
\value{
A \pkg{ggplot2} object of the plotted ROC curve(s) are returned. 
   Use \code{print} function to display the saved object.
}
\description{
Plot the binormal-predicted ROC curve with provided parameters
}
\details{
\code{a} and \code{b} must have the same length. The predicted ROC curve 
   for each \code{a} and \code{b} pair will be plotted.
}
\examples{
binormalPlot <- PlotBinormalFit(c(1, 2), c(0.5, 0.5))
print(binormalPlot)

}
