% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_roots.R
\name{real.roots}
\alias{real.roots}
\title{Test code to exercise testthat}
\usage{
real.roots(a, b, c)
}
\arguments{
\item{a}{The seed variable, default is 123; set to NULL for truly random seed}

\item{b}{The number of non-diseased cases, default is 50}

\item{c}{The number of diseased cases, default is 50}
}
\value{
The roots of the quadratic defined by a, b, c
}
\description{
From website example
}
