% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJafrocGui.R
\name{RJafrocGui}
\alias{RJafrocGui}
\title{Graphical user interface to \pkg{RJafroc} functions}
\usage{
RJafrocGui(useBrowser = FALSE)
}
\arguments{
\item{useBrowser}{a logical variable, if TRUE the default internet browser is used, otherwise RStudio's
internal browser is used. Default is \code{FALSE}. See "Details".}
}
\description{
Start a graphical user interface which is functionally similar to Windows JAFROC software (Version 4.2.1)
}
\details{
For Windows users, we suggest setting \code{useBrowser} to \code{TRUE} due to a bug in RStudio's (Version 0.99.467)
internal broswer which prevents saving plots.
}
\examples{
\dontrun{
## For Windows users:
RJafrocGui(useBrowser = TRUE)

## For other users:
RJafrocGui()
}
}

