% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_erf}
\alias{RSM_erf}
\title{RSM required error function}
\usage{
RSM_erf(x)
}
\arguments{
\item{x}{The value at which to evaluate the function.}
}
\value{
erf
}
\description{
RSM required error function
}
\examples{
RSM_erf(c(-Inf,0.1,0.2,0.3, Inf))

}
