% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_FPF}
\alias{RSM_FPF}
\title{RSM predicted ROC-abscissa as function of z}
\usage{
RSM_FPF(z, lambda)
}
\arguments{
\item{z}{The z-vector at which to evaluate the ROC-abscissa.}

\item{lambda}{The scalar RSM lambda parameter.}
}
\value{
FPF, the abscissa of the ROC
}
\description{
RSM predicted ROC-abscissa as function of z
}
\examples{
RSM_FPF(c(-Inf,0.1,0.2,0.3),1)

}
