% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_NLF}
\alias{RSM_NLF}
\title{RSM predicted FROC abscissa}
\usage{
RSM_NLF(z, lambda)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the FROC abscissa.}

\item{lambda}{The RSM lambda parameter.}
}
\value{
xFROC
}
\description{
RSM predicted FROC abscissa
}
\examples{
RSM_NLF(1,1)

}
