% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_TPF}
\alias{RSM_TPF}
\title{RSM predicted ROC-ordinate as function of z}
\usage{
RSM_TPF(z, mu, lambda, nu, lesDistr)
}
\arguments{
\item{z}{The z-vector at which to evaluate the pdf.}

\item{mu}{The scalar RSM  mu  parameter.}

\item{lambda}{The scalar RSM  lambda parameter.}

\item{nu}{The scalar nu parameter.}

\item{lesDistr}{The lesion distribution 1D vector.}
}
\value{
TPF, the ordinate of the ROC
}
\description{
RSM predicted ROC-ordinate as function of z
}
\examples{
lesDistr <- c(0.1,0.3,0.6)
RSM_TPF(c(-Inf,0.1,0.2,0.3), 1, 1, 0.9, lesDistr)

}
