% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_wLLF_R}
\alias{RSM_wLLF_R}
\title{RSM predicted wAFROC ordinate}
\usage{
RSM_wLLF_R(zeta, mu, nu, lesDistr, relWeights)
}
\arguments{
\item{zeta}{The zeta-vector at which to evaluate the FROC ordinate.}

\item{mu}{The scalar RSM mu parameter.}

\item{nu}{The scalar RSM nu prime parameter.}

\item{lesDistr}{Lesion distribution vector.}

\item{relWeights}{The lesion weights matrix}
}
\value{
wLLF, the ordinate of the wAFROC curve
}
\description{
RSM predicted wAFROC ordinate
}
\examples{
RSM_wLLF_R(c(1,2),1,0.5, lesDistr = c(0.5, 0.4, 0.1), relWeights = c(0.7, 0.2, 0.1)) 


}
