% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_xFROC}
\alias{RSM_xFROC}
\title{RSM predicted FROC abscissa}
\usage{
RSM_xFROC(z, mu, lambda)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the FROC abscissa.}

\item{mu}{The mu parameter of the RSM.}

\item{lambda}{The (intrinsic) lambda parameter of the RSM.}
}
\value{
xFROC
}
\description{
RSM predicted FROC abscissa
}
\examples{
RSM_pdfN(1,1)

}
