% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_xROC}
\alias{RSM_xROC}
\title{RSM predicted ROC-abscissa as function of z}
\usage{
RSM_xROC(z, lambdaP)
}
\arguments{
\item{z}{The z-sample at which to evaluate the ROC-abscissa.}

\item{lambdaP}{The (physical) lambdaP parameter of the RSM.}
}
\value{
xROC, the abscissa of the ROC
}
\description{
RSM predicted ROC-abscissa as function of z
}
\examples{
RSM_xROC(c(-Inf,0.1,0.2,0.3),1)

}
