% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_yFROC}
\alias{RSM_yFROC}
\title{RSM predicted FROC ordinate}
\usage{
RSM_yFROC(z, mu, nu)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the FROC ordinate.}

\item{mu}{The mu parameter of the RSM.}

\item{nu}{The (intrinsic) nu parameter of the RSM.}
}
\value{
yFROC
}
\description{
RSM predicted FROC ordinate
}
\examples{
RSM_yFROC(1,1,0.5)

}
