% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_yROC}
\alias{RSM_yROC}
\title{RSM predicted ROC-ordinate as function of z}
\usage{
RSM_yROC(z, mu, lambdaP, nuP, lesDistr)
}
\arguments{
\item{z}{The value at which to evaluate the pdf.}

\item{mu}{The mu parameter of the RSM.}

\item{lambdaP}{The lambdaP parameter of the RSM.}

\item{nuP}{The nuP parameter of the RSM.}

\item{lesDistr}{The lesion distribution 1D vector.}
}
\value{
yROC
}
\description{
RSM predicted ROC-ordinate as function of z
}
\examples{
lesDistr1D <- c(0.1,0.3,0.6)
RSM_yROC(c(-Inf,0.1,0.2,0.3), 1, 1, 1, lesDistr1D)

}
