% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateFrocDatasetNoSeed.R
\name{SimulateFrocDatasetNoSeed}
\alias{SimulateFrocDatasetNoSeed}
\title{Simulates an MRMC uncorrelated FROC dataset using the RSM - no seed version}
\usage{
SimulateFrocDatasetNoSeed(mu, lambda, nu, zeta1, I, J, K1, K2, perCase)
}
\arguments{
\item{mu}{The mu parameter of the RSM}

\item{lambda}{The intrinsic lambda parameter of the RSM (not the physical parameter)}

\item{nu}{The intrinsic nu parameter of the RSM (not the physical parameter)}

\item{zeta1}{The lowest reporting threshold}

\item{I}{The number of treatments}

\item{J}{The number of readers}

\item{K1}{The number of non-diseased cases}

\item{K2}{The number of diseased cases}

\item{perCase}{A K2 length array containing the numbers of lesions per diseased case}
}
\value{
The return value is an FROC dataset.
}
\description{
Simulates an uncorrelated MRMC FROC dataset for specified numbers of
   readers and treatments
}
\details{
See book chapters on the Radiological Search Model (RSM) for details. 
   In this code correlations between ratings on the same case are assumed to be zero.
}
