% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateRocDataset.R
\name{SimulateRocDataset}
\alias{SimulateRocDataset}
\title{Simulates a binormal model ROC dataset}
\usage{
SimulateRocDataset(I = 1, J = 1, K1, K2, a, deltaA = 0, b, seed = NULL)
}
\arguments{
\item{I}{Number of modalities, default 1}

\item{J}{The number of readers, default 1}

\item{K1}{Number of non-diseased cases}

\item{K2}{Number of diseased cases}

\item{a}{\eqn{a} parameter of binormal model}

\item{deltaA}{Inter-treatment increment in the \eqn{a} parameter, 
default zero}

\item{b}{\eqn{b} parameter of the binormal model}

\item{seed}{Initial seed, default is NULL, for random seed}
}
\value{
An ROC dataset
}
\description{
Simulates an uncorrelated binormal model ROC factorial dataset
}
\details{
See book Chapter 6 for details
}
\examples{
K1 <- 5;K2 <- 7;a <- 1.5;b <- 0.5
rocDataRaw <- SimulateRocDataset(K1 = K1, K2 = K2, a = a, b = b)
  
}
\references{
Chakraborty DP (2017) \emph{Observer Performance Methods for Diagnostic Imaging - Foundations, 
Modeling, and Applications with R-Based Examples}, CRC Press, Boca Raton, FL. 
\url{https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840}
}
