% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsDataset2wAfrocEffectSize.R
\name{SsDataset2wAfrocEffectSize}
\alias{SsDataset2wAfrocEffectSize}
\title{Computing wAFROC effect size scaling factor for an ROC or FROC dataset}
\usage{
SsDataset2wAfrocEffectSize(dataset, lesionHistogram)
}
\arguments{
\item{dataset}{The \strong{pilot} dataset object, assumed to be a NH ROC 
or FROC dataset; if FROC, it is converted to ROC.}

\item{lesionHistogram}{An array containing the fractional distribution of 
number of lesions per diseased case in the \strong{pivotal} FROC study; 
must sum to unity.}
}
\value{
A list containing the linear regression coefficient for model
    lm(eswAfroc ~ -1 + esRoc); the scaling factor by which ROC effect size must be 
    multiplied to get the wAFROC effect size and the R2 of the lm() fit.
}
\description{
Computing wAFROC effect size scaling factor for an ROC or FROC dataset
}
\details{
If dataset is FROC, it is converted to an ROC dataset. The search model 
    is used to fit each treatment-reader combination in the pilot dataset. The median 
    value for each parameter is computed. These are used
    to compute predicted wAFROC and ROC FOMS over a range of values of deltaMu, 
    which are fitted by a straight line constrained to pass throught the origin.
    The slope (scaling factor) and R2 are returned.
}
\examples{
SsDataset2wAfrocEffectSize(dataset02, c(0.7, 0.2, 0.1))
SsDataset2wAfrocEffectSize(DfExtractDataset(dataset04, trts = c(1,2)), c(0.69, 0.2, 0.11))

}
