% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsFrocSampleSize.R
\name{SsFrocSampleSize}
\alias{SsFrocSampleSize}
\title{RSM fitted model for FROC sample size}
\usage{
SsFrocSampleSize(dataset, effectSizeROC, JPivot, KPivot, lesDistr)
}
\arguments{
\item{dataset}{The \strong{pilot} dataset.}

\item{effectSizeROC}{The effect size in ROC-AUC units}

\item{JPivot}{The number of readers in the pivotal study}

\item{KPivot}{The number of cases in the pivotal study}

\item{lesDistr}{A 1D array containing the probability mass function of
number of lesions per diseased case in the \strong{pivotal FROC}
study.}
}
\value{
A list containing:
   \itemize{
   \item \code{effectSizeROC}, the specified ROC effect size.
   \item \code{scaleFactor}, the factor by which the ROC effect size
   must by multiplied to get the wAFROC effect size.
   \item \code{powerRoc}, the ROC power.
   \item \code{powerFroc}, the wAFROC power.
   
}
}
\description{
RSM fitted model for FROC sample size
}
\details{
See \url{https://dpc10ster.github.io/RJafrocQuickStart/froc-sample-size.html}
   for explanation of the FROC sample size estimation procedure.
}
\examples{

\donttest{
## Examples with CPU or elapsed time > 5s
## user system elapsed
## SsFrocSampleSize 8.102  0.023   8.135

## SsFrocSampleSize(DfExtractDataset(dataset04, trts = c(1,2)), 
## effectSizeROC = 0.03, JPivot = 5, KPivot = 100, lesDistr = c(0.69, 0.2, 0.11))
}

}
