% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StORHAnalysisSp.R
\name{StORHAnalysisSp}
\alias{StORHAnalysisSp}
\title{Significance testing using Obuchowski-Rockette variance components for split plot datasets}
\usage{
StORHAnalysisSp(dataset, FOM, FPFValue, alpha = 0.05, analysisOption = "ALL")
}
\arguments{
\item{dataset}{The split plot dataset object, types A and C are currently supported}

\item{FOM}{The figure of merit}

\item{FPFValue}{Only needed for \code{LROC} data \strong{and} FOM = "PCL" or "ALROC";
where to evaluate a partial curve based figure of merit. The default is 0.2.}
}
\value{
A list object containing the following \code{data.frames}: 
    \itemize{
    \item{\code{foms}}: the figures of merit for different treatment-reader combinations 
    \item{\code{TRanova}}: the OR treatment-reader ANOVA table 
    \item{\code{VarCom}}: the OR variance-components \code{Cov1}, \code{Cov2}, 
    \code{Cov3}, \code{Var} and correlations \code{rho1}, \code{rho2} and \code{rho3} 
    \item{\code{IndividualTrt}}: the individual treatment mean-squares, \code{Var} and \code{Cov2} values
    \item{\code{IndividualRdr}}: the individual reader mean-squares, \code{Var} and \code{Cov1} values
    }
}
\description{
Significance testing using Obuchowski-Rockette variance components for split plot datasets
}
\details{
The variance components are obtained using \link{StSignificanceTesting} 
    with \code{method = "OR"}.
}
\examples{
  
}
