% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSP.R
\name{StSP}
\alias{StSP}
\title{Perform significance testing for split-plot datasets}
\usage{
StSP(dataset, FOM, alpha = 0.05, analysisOption = "RRRC")
}
\arguments{
\item{dataset}{The dataset to be analyzed, see \code{\link{RJafroc-package}}. 
The dataset design must be "SPLIT-PLOT-A" or "SPLIT-PLOT-C" (see Note).}

\item{FOM}{The figure of merit}

\item{alpha}{The significance level of the test, default is 0.05}

\item{analysisOption}{Factors regarded as random vs. those regarded as fixed:
\itemize{ 
   \item \code{"RRRC"} = random-reader random case, the default,
   \item \code{"FRRC"} = fixed-reader random case, 
   \item \code{"RRFC"} = random-reader fixed case, 
}}
}
\value{
Results of the analysis
}
\description{
Performs Obuchowski-Rockette significance testing for 
   split-plot datasets as described in Hillis 2014.
}
\note{
Two split-plot designs are supported; these are described in 
     Table VII in Hillis 2014:
\itemize{ 
   \item SPLIT-PLOT-A = {Reader nested within test, labeled (a) in Table VII;
   see for example \code{rocSpA.xlsx}}
   \item SPLIT-PLOT-C = {Case nested within reader, labeled (c) in Table VII;
   see for example \code{rocSpC.xlsx}}
}
The included toy ROC datasets are in directory \code{inst/extdata/toyFiles/ROC}.
}
\examples{
## Analyze included dataset 
fileName <- system.file("extdata", "/toyFiles/ROC/rocSpA.xlsx", 
package = "RJafroc", mustWork = TRUE)
dsSpA <- DfReadSP(fileName)
ret <- StSP(dsSpA, FOM = "Wilcoxon")

## Analyze embedded dataset 
ret <- StSP(datasetFROCSpC, FOM = "wAFROC")


}
\references{
Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader
multicase radiological imaging data, Statistics in medicine 33, 330-360.
}
