% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCpp.R
\name{testCpp}
\alias{testCpp}
\title{Testing cpp code TBA!!}
\usage{
testCpp(mu, nu)
}
\arguments{
\item{mu}{The mu parameter.}

\item{nu}{The nu parameter.}
}
\description{
test cpp code, temporary function, move to tests as they are done and finally delete this function
}
\keyword{internal}
