% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilAnalyticalAucsRSM.R
\name{UtilAnalyticalAucsRSM}
\alias{UtilAnalyticalAucsRSM}
\title{RSM ROC/AFROC/wAFROC AUC calculator}
\usage{
UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = -Inf, lesDistr, relWeights = 0)
}
\arguments{
\item{mu}{The mean of the Gaussian distribution for the 
ratings of latent LLs (continuous ratings of lesions that 
are found by the search mechanism).}

\item{lambda}{The \emph{intrinsic} Poisson distribution parameter, 
which describes 
the random number of latent NLs (suspicious regions that do not 
correspond to actual lesions) per case.}

\item{nu}{The \emph{intrinsic} \code{nu} parameters, 
the success probability of the binomial distribution describing 
the random numbers of latent LLs (suspicious regions that correspond 
to actual lesions) per diseased case.}

\item{zeta1}{The lowest reporting threshold, the default is \code{-Inf}.}

\item{lesDistr}{See \code{\link{PlotRsmOperatingCharacteristics}}.}

\item{relWeights}{The relative weights of the lesions; a vector of 
length \code{maxLL}; if zero, the default, equal weights are assumed.}
}
\value{
A list containing the ROC, AFROC and wAFROC AUCs corresponding to the 
   specified parameters
}
\description{
{Returns the ROC, AFROC and wAFROC AUCs corresponding to 
   specified RSM parameters. See also \code{\link{UtilAucPROPROC}}, 
   \code{\link{UtilAucBinormal}} and \code{\link{UtilAucCBM}}}
}
\examples{
mu <- 1;lambda <- 1;nu <- 1
lesDistr <- rbind(c(1, 0.9), c(2, 0.1)) 
## i.e., 90\% of dis. cases have one lesion, and 10\% have two lesions

UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = -Inf, lesDistr)

relWeights <- c(0.05, 0.95)
UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = 0, lesDistr, relWeights)


}
\references{
Chakraborty DP (2017) \emph{Observer Performance Methods for Diagnostic Imaging - Foundations, 
Modeling, and Applications with R-Based Examples}, CRC Press, Boca Raton, FL. 
\url{https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840}

Chakraborty DP (2006) A search model and figure of merit for observer data acquired according to the free-response 
paradigm, Phys Med Biol 51, 3449-3462.

Chakraborty DP (2006) ROC Curves predicted by a model of visual search, Phys Med Biol 51, 3463--3482.
}
