% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilAucPROPROC.R
\name{UtilAucPROPROC}
\alias{UtilAucPROPROC}
\title{PROPROC AUC function}
\usage{
UtilAucPROPROC(c1, da)
}
\arguments{
\item{c1}{The c-parameter of the PROPROC model, since \strong{c is a reserved 
function in R}.}

\item{da}{The da-parameter of the PROPROC model.}
}
\value{
PROPROC model-predicted ROC-AUC for the specified parameters
}
\description{
{Returns the PROPROC ROC-AUC corresponding to specified 
   parameters. See also \code{\link{UtilAucsRSM}}, \code{\link{UtilAucBinormal}} 
   and \code{\link{UtilAucCBM}}}
}
\examples{
c1 <- .2;da <- 1.5
UtilAucPROPROC(c1,da)


}
\references{
Metz CE, Pan X (1999) Proper Binormal ROC Curves: Theory and Maximum-Likelihood Estimation, J Math Psychol 43(1):1-33.
}
