% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilDBM2ORVarComp.R
\name{UtilDBM2ORVarComp}
\alias{UtilDBM2ORVarComp}
\title{Convert from DBM to OR variance components}
\usage{
UtilDBM2ORVarComp(K, DBMVarComp)
}
\arguments{
\item{K}{Total number of cases}

\item{DBMVarComp}{DBM variance components, a data.frame 
containing VarR, VarC, VarTR, VarTC, VarRC and VarErr}
}
\value{
\code{UtilDBM2ORVarComp} returns the equivalent OR Variance components
}
\description{
\code{UtilDBM2ORVarComp} converts from DBM variance components to OR 
   variance components
}
\examples{
DBMVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "DBMH")$ANOVA$VarCom
UtilDBM2ORVarComp(114, DBMVarCom)

ORVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "ORH")$ANOVA$VarCom
UtilOR2DBMVarComp(114, ORVarCom)


}
