% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFigureOfMeritX.R
\name{UtilFigureOfMeritX}
\alias{UtilFigureOfMeritX}
\title{Empirical figures of merit for cross-modality dataset}
\usage{
UtilFigureOfMeritX(dsX, FOM = "wAFROC")
}
\arguments{
\item{dsX}{The cross-modality dataset to be analyzed.}

\item{FOM}{The figure of merit; the default is \code{"wAFROC"}}
}
\value{
Two data frames: 
* \code{c(I2, J)} data frame, FOMs averaged over the first modality, where the row names are modality IDS of the 
   second modality 
* \code{c(I1, J)} data frames, FOMs averaged over the second modality, where the row names are modality IDs of the 
   first modality, 
* The column names are the \code{readerID}'s.
}
\description{
Calculate specified empirical figure of merit
   for each modality-reader combination in a cross-modality dataset
}
\details{
The allowed FOMs depend on the \code{dataType} field of the 
   \code{dataset} object.  See \code{\link{UtilFigureOfMerit}}.
}
\examples{

##UtilFigureOfMeritX(datasetXModality, FOM = "wAFROC")


}
\references{
Chakraborty DP (2017) \emph{Observer Performance Methods for Diagnostic Imaging - Foundations, 
Modeling, and Applications with R-Based Examples}, CRC Press, Boca Raton, FL. 
\url{https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840}

Chakraborty DP, Berbaum KS (2004) Observer studies involving detection and localization: modeling, analysis, and validation, 
Medical Physics, 31(8), 1--18.

Song T, Bandos AI, Rockette HE, Gur D (2008) On comparing methods for discriminating between actually negative and actually positive subjects 
with FROC type data, Medical Physics 35 1547--1558.

Popescu LM (2011) Nonparametric signal detectability evaluation using an exponential transformation of the FROC curve, 
Medical Physics, 38(10), 5690. 

Obuchowski NA, Lieber ML, Powell KA (2000) Data Analysis for Detection and Localization of Multiple Abnormalities 
with Application to Mammography, Acad Radiol, 7:7 553--554.

Swensson RG (1996) Unified measurement of observer performance in detecting and localizing target objects on images, 
Med Phys 23:10, 1709--1725.
}
