% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilLesionWeightsDistr.R
\name{UtilLesionWeightsDistr}
\alias{UtilLesionWeightsDistr}
\title{Lesion weights distribution}
\usage{
UtilLesionWeightsDistr(dataset)
}
\arguments{
\item{dataset}{The supplied dataset.}
}
\value{
Two characteristics of an FROC dataset, apart from the 
   ratings, affect the FOM: the distribution of lesion per case and the 
   distribution of lesion weights. This function addresses the weights. 
   The distribution of lesions is addressed in \link{UtilLesionDistr}. 
   \code{lesWghtDistr} is an [1:nRow,1:(maxLL+1)] array, where 
   \code{nRow} is the number of \bold{unique} values of lesions per case 
   in the dataset. The first column enumerates the number of lesions per 
   case, while the remaining columns contain the weights (equal weighting assumed). 
   Missing values are filled with \code{-Inf}. This parameter is not to be 
   confused with the \code{lesionWeight} list member in an FROC dataset, 
   which enumerates the weights of lesions on \bold{individual} cases. See 
   \link{PlotRsmOperatingCharacteristics} for a function that depends on 
   \code{lesWghtDistr}. See Chapter00Vignette2 for a fuller explanation.
}
\description{
The lesion weights  distribution, assuming equal weights. 
   Needed to plot empirical curves for FROC data (ROC, wAFROC).
}
\examples{
UtilLesionWeightsDistr (dataset11) # FROC data
UtilLesionWeightsDistr (dataset02) # ROC data
UtilLesionWeightsDistr (datasetCadLroc) # LROC data

}
