% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilOR2DBMVarComp.R
\name{UtilOR2DBMVarComp}
\alias{UtilOR2DBMVarComp}
\title{Convert from OR to DBM variance components}
\usage{
UtilOR2DBMVarComp(K, ORVarComp)
}
\arguments{
\item{K}{Total number of cases}

\item{ORVarComp}{OR variance components, a data.frame 
containing VarR, VarTR, Cov1, Cov2, Cov3 and Var}
}
\value{
\code{UtilOR2DBMVarComp} returns the equivalent DBM variance components
}
\description{
\code{UtilOR2DBMVarComp} converts from OR to DBM variance components.
}
\examples{
DBMVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "DBMH")$ANOVA$VarCom
UtilDBM2ORVarComp(114, DBMVarCom)

ORVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "ORH")$ANOVA$VarCom
UtilOR2DBMVarComp(114, ORVarCom)



}
