% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilORVarComp.R
\name{UtilORVarComp}
\alias{UtilORVarComp}
\title{Obuchowski-Rockette variance components for dataset}
\usage{
UtilORVarComp(
  dataset,
  FOM,
  covEstMethod = "jackknife",
  FPFValue = 0.2,
  nBoots = 200,
  seed = NULL
)
}
\arguments{
\item{dataset}{Factorial one-treatment or cross-modality two-treatment dataset}

\item{FOM}{Figure of merit}

\item{covEstMethod}{The covariance estimation method, "jackknife" 
(the default) or "bootstrap" or "DeLong" ("DeLong" is applicable only for 
FOM = "Wilcoxon").}

\item{FPFValue}{Only needed for \code{LROC} data \strong{and} FOM = "PCL" 
or "ALROC": the \code{FPFValue} at which to evaluate a partial curve 
based figure of merit. The default is 0.2.}

\item{nBoots}{The number of bootstraps (default 200).Only needed for 
covEstMethod = "bootstrap".}

\item{seed}{Only needed for the bootstrap covariance estimation method. 
The initial seed for the random number generator, the default is 
\code{NULL}, for random seed.}
}
\value{
A list containing the following \code{data.frames}: 
    \itemize{
    \item{\code{foms}}: the figures of merit for different modality-reader combinations 
    \item{\code{TRanova}}: the OR modality-reader ANOVA table 
    \item{\code{VarCom}}: the OR variance-components \code{Cov1}, \code{Cov2}, 
    \code{Cov3}, \code{Var} and correlations \code{rho1}, \code{rho2} and \code{rho3} 
    \item{\code{IndividualTrt}}: the individual modality mean-squares, \code{Var} and \code{Cov2} values
    \item{\code{IndividualRdr}}: the individual reader mean-squares, \code{Var} and \code{Cov1} values
    }
}
\description{
Obuchowski-Rockette variance components for dataset
}
\details{
The variance components are identical to those obtained using 
    \link{St} with \code{method = "OR"}.
}
\examples{
## use the default jackknife for covEstMethod
vc <- UtilORVarComp(dataset02, FOM = "Wilcoxon")

##UtilORVarComp(dataset02, FOM = "Wilcoxon", covEstMethod = "bootstrap", 
##nBoots = 2000, seed = 100)$VarCom 

##UtilORVarComp(dataset02, FOM = "Wilcoxon", covEstMethod = "DeLong")$VarCom

vc <- UtilORVarComp(datasetXModality, FOM = "wAFROC") 
  
}
