% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilSpecifyLesionWeightsDistr.R
\name{UtilSpecifyLesionWeightsDistr}
\alias{UtilSpecifyLesionWeightsDistr}
\title{Specify lesion weight distribution array}
\usage{
UtilSpecifyLesionWeightsDistr(maxLL, relWeights = 0)
}
\arguments{
\item{maxLL}{The maximum number of lesions per case in the dataset.}

\item{relWeights}{The relative weights of the lesions; a vector of 
length maxLL; if zero, the default, equal weights are assumed.}
}
\value{
lesWghtDistr The lesion weights distribution array
}
\description{
{specify the lesion weight distribution array}
}
\examples{
 



}
