% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilVarComponentsOR.R
\name{UtilVarComponentsDBM}
\alias{UtilVarComponentsDBM}
\title{Utility for Dorfman-Berbaum-Metz variance components}
\usage{
UtilVarComponentsDBM(dataset, FOM, FPFValue = 0.2)
}
\arguments{
\item{dataset}{The dataset object}

\item{FOM}{The figure of merit}

\item{FPFValue}{Only needed for \code{LROC} data \strong{and} FOM = "PCL" or "ALROC";
where to evaluate a partial curve based figure of merit. The default is 0.2.}
}
\value{
A list object containing the variance components.
}
\description{
Utility for Dorfman-Berbaum-Metz variance components
}
\examples{
UtilVarComponentsDBM(dataset02, FOM = "Wilcoxon")

  
}
