% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilVarComponentsORX.R
\name{UtilVarComponentsORX}
\alias{UtilVarComponentsORX}
\title{Utility for estimating Obuchowski-Rockette variance components for crossed modality factorial datasets}
\usage{
UtilVarComponentsORX(dsX, FOM, FPFValue = 0.2)
}
\arguments{
\item{dsX}{The crossed modality factorial dsX object}

\item{FOM}{The figure of merit}
}
\value{
A list containing the following \code{data.frames}: 
    \itemize{
    \item{\code{foms}}: the figures of merit for different modality-reader combinations 
    \item{\code{TRanova1}}: the OR modality-reader ANOVA table 
    \item{\code{VarCom}}: the OR variance-components \code{Cov1}, \code{Cov2}, 
    \code{Cov3}, \code{Var} and correlations \code{rho1}, \code{rho2} and \code{rho3} 
    \item{\code{IndividualTrt}}: the individual modality mean-squares, \code{Var} and \code{Cov2} values
    \item{\code{IndividualRdr}}: the individual reader mean-squares, \code{Var} and \code{Cov1} values
    }
}
\description{
Utility for estimating Obuchowski-Rockette variance components for crossed modality factorial datasets
}
\examples{
## use the default jackknife for covEstMethod
## vc <- UtilVarComponentsORX(datasetXModality, FOM = "Wilcoxon")
## str(vc) 

}
