% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChisqrGoodnessOfFit.R
\name{ChisqrGoodnessOfFit}
\alias{ChisqrGoodnessOfFit}
\title{Compute the chisquare goodness of fit statistic for specified ROC data fitting model}
\usage{
ChisqrGoodnessOfFit(zetas, lesDistr, fpCounts, tpCounts, parameters, model)
}
\arguments{
\item{zetas}{The estimated thresholds of the fitting model}

\item{lesDistr}{The lesion distribution matrix; \code{NA} for "BINORMAL" and "CBM}

\item{fpCounts}{The FP counts table}

\item{tpCounts}{The TP counts table}

\item{parameters}{The parameters of the model, see details}

\item{model}{The fitting model: "BINORMAL", "CBM" or "RSM}
}
\value{
The return value is a list with the following elements:

\item{chisq}{The chi-square statistic}

\item{pVal}{The p-value of the fit}

\item{df}{The degrees of freedom}
}
\description{
Compute the chisquare goodness of fit statistic for specified ROC data fitting model
}
\details{
For model = "BINORMAL" the parameters are c(a,b).
For model = "CBM" the parameters are c(mu,alpha).
For model = "RSM" the parameters are c(mu,lambdaP, nuP).
For conventional ROC models, lesDistr should be \code{NULL}.
}
