% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset01}
\alias{dataset01}
\title{TONY FROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:2, 1:5, 1:185, 1:3], ratings of non-lesion localizations, NLs}
\item{\code{LL}}{, num [1:2, 1:5, 1:89, 1:2], ratings of lesion localizations, LLs}
\item{\code{lesionVector}}{, int [1:89], number of lesions per diseased case}
\item{\code{lesionID}}{, num [1:89, 1:2], numeric labels of lesions on diseased cases}
\item{\code{lesionWeight}}{, num [1:89, 1:2], weights (or clinical importances) of lesions}
\item{\code{dataType}}{, chr "FROC", the data type}
\item{\code{modalityID}}{, chr [1:2] "BT" "DM", treatment labels}
\item{\code{readerID}}{, chr [1:5] "1" "2" "3" "4" ..., reader labels}
\item{\code{design}}{, chr "CROSSED", study design}
\item{\code{normalCases}}{, int [1:96] 1 4 7 8 9 10 15 17 18 20 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:89] 2 3 5 6 11 12 13 16 19 24 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:2, 1:5, 1:185, 1:4] 1 1 1 1 ..., truth table structure}
\item{\code{datasetName}}{, chr "TONY"}
}
}
\usage{
dataset01
}
\description{
This is referred to in the book as the "TONY" dataset. It consists of 185 cases, 
89 of which are diseased, interpreted in two treatments
("BT" = breast tomosynthesis and "DM" = digital mammography) by five radiologists using the FROC paradigm.
}
\examples{
str(dataset01)
PlotEmpiricalOperatingCharacteristics(dataset = dataset01, opChType = "wAFROC")$Plot

}
\references{
Chakraborty DP, Svahn T (2011) Estimating the parameters of a model 
of visual search from ROC data: an alternate method for fitting proper ROC curves. 
PROC SPIE 7966.
}
\keyword{datasets}
