% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset08}
\alias{dataset08}
\title{Monica Penedo ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:5, 1:5, 1:112, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:5, 1:5, 1:64, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:64], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:64, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:64, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:5] "0" "1", treatment labels}
\item{\code{readerID}}{, chr [1:5] "0" "1" "2" ..., reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, nt [1:48] 1 2 3 4 5 6 7 8 9 10..., normal case labels}
\item{\code{abnormalCases}}{, int [1:64] 49 50 51 52 53 54 55 56 57 58 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:5, 1:5, 1:112, 1:2] 1 1 1 1 1 1 ..., truth table structure}
\item{\code{datasetName}}{, chr "Penedo"}
}
}
\usage{
dataset08
}
\description{
This is referred to in the book as the "PEN" dataset. It consists of 112 cases, 
64 of which are diseased, interpreted in five treatments (basically different image compression
algorithms) by five 
radiologists using the FROC paradigm (the inferred ROC dataset is included; the original FROC data
is lost).
}
\examples{
str(dataset08)
PlotEmpiricalOperatingCharacteristics(dataset = dataset08, opChType = "ROC")$Plot


}
\references{
Penedo et al. Free-Response Receiver Operating Characteristic 
Evaluation of Lossy JPEG2000 and Object-based Set Partitioning in 
Hierarchical Trees Compression of Digitized Mammograms. 
Radiology. 2005;237(2):450-457.
}
\keyword{datasets}
