% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset10}
\alias{dataset10}
\title{Marc Ruschin ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:3, 1:8, 1:90, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:3, 1:8, 1:40, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:40], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:40, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:40, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:3] "1" "2" "3", treatment labels}
\item{\code{readerID}}{, chr [1:8] "1" "2" ..., reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:50] 1 2 3 4 5 6 7 8 9 10 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:40] 51 52 53 54 55 56 57 58 59 60 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:3, 1:8, 1:90, 1:2] 1 1 1 1 1 1 ..., truth table structure}
\item{\code{datasetName}}{"Ruschin"}
}
}
\usage{
dataset10
}
\description{
This is referred to in the book as the "RUS" dataset. It consists of 90 cases, 
40 of which are diseased, the images were
acquired at three dose levels, which can be regarded as treatments.
"0" = conventional film radiographs, "1" = digitized  images viewed on monitors, Eight 
radiologists interpreted the cases using the FROC paradigm. These have been reduced to 
ROC data by using the highest ratings (the original FROC data is lost).
}
\examples{
str(dataset10)
PlotEmpiricalOperatingCharacteristics(dataset = dataset10, opChType = "ROC")$Plot


}
\references{
Ruschin M, et al. Dose dependence of mass and microcalcification 
detection in digital mammography: free response human observer studies. 
Med Phys. 2007;34:400 - 407.
}
\keyword{datasets}
