% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset11}
\alias{dataset11}
\title{Dobbins 1 FROC dataset}
\format{A list with 8 elements:
\itemize{
\item{\code{NL}}{, num [1:4, 1:5, 1:158, 1:4], ratings of non-lesion localizations, NLs}
\item{\code{LL}}{, num [1:4, 1:5, 1:115, 1:20], ratings of lesion localizations, LLs}
\item{\code{lesionVector}}{, int [1:115], number of lesions per diseased case}
\item{\code{lesionID}}{, num [1:115, 20], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:115, 20] weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "FROC", the data type}
\item{\code{modalityID}}{, chr [1:4] "1" "2" ..., treatment labels}
\item{\code{readerID}}{, chr [1:5] "1" "2" ..., reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:43] 6 9 14 27 62 66 70 71 83 91 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:115] 1 2 3 5 7 8 10 11 13 17 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:4, 1:5, 1:158, 1:21] 1 1 1 1 1 1 ..., truth table structure}
}}
\usage{
dataset11
}
\description{
This is referred to in the book as the "DOB1" dataset. Dobbins et al conducted a 
multi-institutional, MRMC study to compare the performance of digital tomosynthesis 
(GE's VolumeRad device), dual-energy (DE) imaging, and conventional chest 
radiography for pulmonary nodule detection and management. 
All study images were obtained with a flat-panel detector developed by GE. 
The case set consisted of 158 subjects, of which 43 were non-diseased and 
the rest had 1 - 20 pulmonary nodules independently verified, using with CT 
images, by 3 experts who did not participate in the observer study. The 
study used FROC paradigm data collection. There are 
4 treatments labeled 1 - 4 (conventional chest x-ray, CXR, CXR augmented
with dual-energy (CXR+DE), VolumeRad digital tomosynthesis images and 
VolumeRad augmented with DE (VolumeRad+DE).
}
\examples{
str(dataset11)


}
\references{
Dobbins III JT et al. Multi-Institutional Evaluation of Digital 
Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography 
for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.
}
\keyword{datasets}
