% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset12}
\alias{dataset12}
\title{Dobbins 2 ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:4, 1:5, 1:152, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:4, 1:5, 1:88, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:88], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:88, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:88, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:4] "1" "2" "3" "4", treatment labels}
\item{\code{readerID}}{, chr [1:5] "1" "2" "3" "4" ..., reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:64] 6 9 14 20 27 33 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:88] 1 2 3 5 7 8 10 11 13 18 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:4, 1:5, 1:152, 1:2] 1 1 1 1 1 1 ..., truth table structure}
}
}
\usage{
dataset12
}
\description{
This is referred to in the code as the "DOB2" dataset. It contains actionability
ratings, i.e., do you recommend further follow up on the patient, one a 1 (definitely not)
to 5 (definitely yes), effectively an ROC dataset using a 5-point rating scale.
}
\examples{
str(dataset11)


}
\references{
Dobbins III JT et al. Multi-Institutional Evaluation of Digital 
Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography 
for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.
}
\keyword{datasets}
