% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset12}
\alias{dataset12}
\title{Dobbins 2 ROC dataset}
\format{A list with 8 elements:
\itemize{
\item{\code{NL}}{ Ratings array [1:4, 1:5, 1:152, 1], of false positives, FPs}
\item{\code{LL}}{ Ratings array [1:4, 1:5, 1:88, 1], of true positives, TPs}
\item{\code{lesionVector}}{ array [1:88], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{ array [1:88, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{ array [1:88, 1], weights (or clinical importance) of lesions, all set to 1}
\item{\code{dataType}}{ "ROC", the data type}
\item{\code{modalityID}}{ [1:2] "0" "1", treatment labels}
\item{\code{readerID}}{ [1:4] "0" "1" "2" ..., reader labels}
}}
\usage{
dataset12
}
\description{
This is referred to in the code as the "DOB2" dataset. It contains actionability
ratings, i.e., do you recommend further follow up on the patient, one a 1 (definitely not)
to 5 (definitely yes), effectively an ROC dataset using a 5-point rating scale.
}
\examples{
str(dataset11)


}
\references{
Dobbins III JT et al. Multi-Institutional Evaluation of Digital 
Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography 
for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.
}
\keyword{datasets}
