% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset14}
\alias{dataset14}
\title{Federica Zanca real (as opposed to inferred) ROC dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1:2, 1:4, 1:200, 1], ratings of false positives, FPs}
\item{\code{LL}}{, num [1:2, 1:4, 1:100, 1], ratings of true positives, TPs}
\item{\code{lesionVector}}{, int [1:100], number of lesions per diseased case, all set to 1}
\item{\code{lesionID}}{, num [1:100, 1], labels of lesions on diseased cases, all set to 1}
\item{\code{lesionWeight}}{, num [1:100, 1], weights (or clinical importances) of lesions, all set to 1}
\item{\code{dataType}}{, chr "ROC", the data type}
\item{\code{modalityID}}{, chr [1:2] "4" "5", treatment labels}
\item{\code{readerID}}{, chr [1:4] "1" "2" "3" "4", reader labels}
\item{\code{design}}{, "CROSSED", study design}
\item{\code{normalCases}}{, int [1:100] 100 101 102 103 104 105 106 ..., normal case labels}
\item{\code{abnormalCases}}{, int [1:100] 0 1 2 3 4 5 6 7 8 9 ..., abnormal case labels}
\item{\code{truthTableStr}}{, num [1:2, 1:4, 1:200, 1:2] 1 1 1 1 1 1 ..., truth table structure}
\item{\code{datasetName}}{, chr "FedericaRealRoc"}
}
}
\usage{
dataset14
}
\description{
This is referred to in the book as the "FZR" dataset. It is a real ROC study, 
conducted on the same images and using the same radiologists, on treatments
"4" and "5" of dataset04. This was compared to highest rating inferred ROC
data from dataset04 to conclude, erroneously, that the highest rating assumption
is invalid. See book Section 13.6.2.
}
\examples{
str(dataset14)


}
\references{
Zanca F, Hillis SL, Claus F, et al (2012) Correlation of free-response and 
receiver-operating-characteristic area-under-the-curve estimates: Results from 
independently conducted FROC/ROC studies in mammography. 
Med Phys. 39(10):5917-5929.
}
\keyword{datasets}
