% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasetCadSimuFroc}
\alias{datasetCadSimuFroc}
\title{Simulated FROC CAD vs. RAD dataset}
\format{
A list with 12 elements:
\itemize{
\item{\code{NL}}{, num [1, 1:10, 1:200, 1], ratings of non-lesion localizations, NLs}
\item{\code{LL}}{, num [1, 1:10, 1:80, 1], ratings of lesion localizations, LLs}
\item{\code{lesionVector}}{, int [1:80], number of lesions per diseased case, all set to one}
\item{\code{lesionID}}{, num [1:80, 1], lesions labels on diseased cases, all set to one}
\item{\code{lesionWeight}}{, num [1:80, 1], weights, all set to one}
\item{\code{dataType}}{, chr "FROC", the data type}
\item{\code{modalityID}}{, chr "1", treatment label}
\item{\code{readerID}}{, chr [1:10] "1" "2",  reader labels}
\item{\code{datasetName}}{, chr "dataset Cad Simulated Froc"}
}
}
\usage{
datasetCadSimuFroc
}
\description{
Simulated FROC CAD vs. RAD dataset suitable for checking code. It was generated 
   from datasetCadLroc using SimulateFrocFromLrocData.R.
}
\examples{
str(datasetCadSimuFroc)

}
\keyword{datasets}
