% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasetCadSimuFroc}
\alias{datasetCadSimuFroc}
\title{Simulated FROC CAD vs. RAD dataset}
\format{A list with 8 elements:
\itemize{
\item{\code{NL}}{ Ratings array [1, 1:10, 1:200, 1], of non-lesion localizations, NLs}
\item{\code{LL}}{ Ratings array [1, 1:10, 1:80, 1], of lesion localizations, LLs}
\item{\code{lesionVector}}{ array [1:80], number of lesions per diseased case, all set to one}
\item{\code{lesionID}}{ array [1:80, 1], lesions labels on diseased cases, all set to one}
\item{\code{lesionWeight}}{ array [1:80, 1], weights, all set to one}
\item{\code{dataType}}{ "FROC", the data type}
\item{\code{modalityID}}{ "1", treatment label}
\item{\code{readerID}}{ [1:10] "1" "2",  reader labels}
}}
\usage{
datasetCadSimuFroc
}
\description{
Simulated FROC CAD vs. RAD dataset suitable for checking code. It was generated 
   from datasetCadLroc using SimulateFrocFromLrocData.R.
}
\examples{
str(datasetCadSimuFroc)

}
\keyword{datasets}
