% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasetFROCSpC}
\alias{datasetFROCSpC}
\title{#' Simulated FROC SPLIT-PLOT-C dataset}
\format{
A list with 3 elements: \code{$ratings}, \code{$lesions} and \code{$descriptions}; \code{$ratings}
   contain 3 elements, \code{$NL}, \code{$LL} and \code{$LL_IL} as sub-lists; \code{$lesions}
   contain 3 elements, \code{$perCase}, \code{$IDs} and \code{$weights} as sub-lists; \code{$descriptions}
   contain 7 elements, \code{$fileName}, \code{$type}, \code{$name},
   \code{$truthTableStr}, \code{$design}, \code{$modalityID} and \code{$readerID} as sub-lists;
\itemize{
\item\code{rating$NL} num Array[1:2, 1:4, 1:200, 1:7] ratings of non-lesion localizations NLs
\item\code{rating$LL} num Array[1:2, 1:4, 1:100, 1:3] ratings of lesion localizations LLs
\item\code{rating$LL_IL} NA This placeholder is used only for LROC data
\item\code{lesions$perCase} int Array[1:100], number of lesions per diseased case
\item\code{lesions$IDs} num Array[1:100, 1:3] , numeric labels of lesions on diseased cases
\item\code{lesions$weights} num Array[1:100, 1:3], weights (or clinical importance) of lesions
\item\code{descriptions$fileName} chr "datasetFROCSpC", base name of dataset in `data` folder
\item\code{descriptions$type} chr "FROC" the data type
\item\code{descriptions$name} chr "SIM-FROC-SPLIT-PLOT-C", the name of the dataset
\item\code{descriptions$truthTableStr} NA truth table structure
\item\code{descriptions$design} chr "FCTRL-X-MOD", study design factorial dataset
\item\code{descriptions$modalityID} chr Array[1:2] treatment label(s)
\item\code{descriptions$readerID} chr Array[1:4] reader labels
}
}
\usage{
datasetFROCSpC
}
\description{
#' Simulated FROC SPLIT-PLOT-C dataset
}
\keyword{datasets}
