% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfDatasetLroc2Roc.R
\name{DfDatasetLroc2Roc}
\alias{DfDatasetLroc2Roc}
\title{Convert an LROC dataset to a ROC dataset}
\usage{
DfDatasetLroc2Roc(dataset)
}
\arguments{
\item{dataset}{The \strong{LROC} dataset to be converted.}
}
\value{
An ROC dataset
}
\description{
Converts an LROC dataset to an ROC dataset
}
\details{
The conversion is effected by taking the maximum rating on each diseased case, 
   which could be a TPCl (true positive correct localization) or a TPIl 
   (true positive incorrect localization) rating, whichever has the higher rating.
}
\examples{
str(datasetCadLroc)
rocDataSet <- DfDatasetLroc2Roc(datasetCadLroc)
str(rocDataSet)


}
