% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfLroc2Roc.R
\name{DfLroc2Roc}
\alias{DfLroc2Roc}
\title{Convert an LROC dataset to a ROC dataset}
\usage{
DfLroc2Roc(dataset)
}
\arguments{
\item{dataset}{The \strong{LROC} dataset to be converted.}
}
\value{
An ROC dataset
}
\description{
Converts an LROC dataset to an ROC dataset
}
\details{
For the diseased cases one takes the maximum rating on each diseased case, 
   which could be a LLCl ("true positive" correct localization) or a LLIl 
   ("true positive" incorrect localization) rating, whichever has the higher rating.
   For non-diseased cases the NL arrays are identical.
}
\examples{
str(datasetCadLroc)
rocDataSet <- DfLroc2Roc(datasetCadLroc)
str(rocDataSet)


}
