% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSP.R
\name{DfReadSP}
\alias{DfReadSP}
\title{Read a SPLIT PLOT data file (not factorial)}
\usage{
DfReadSP(fileName)
}
\arguments{
\item{fileName}{A string specifying the name of the file.}
}
\value{
A dataset with the structure specified in \code{\link{RJafroc-package}}.
}
\description{
Read a disk file and create an ROC or FROC dataset object
}
\examples{
fileName <- system.file("extdata", "toyFiles/ROC/rocCr.xlsx", 
package = "RJafroc", mustWork = TRUE)
ds <- DfReadSP(fileName)


}
