% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSP.R
\name{DFReadSP}
\alias{DFReadSP}
\title{Read a SPLIT PLOT data file}
\usage{
DFReadSP(fileName)
}
\arguments{
\item{fileName}{A string specifying the name of the file. 
The file-extension must match the format specified below.}
}
\value{
A dataset with the structure specified in \code{\link{RJafroc-package}}.
}
\description{
Read a disk file and create a ROC or FROC dataset object 
   from it.
}
\examples{
fileName <- system.file("extdata", "toyFiles/ROC/rocCr.xlsx", 
package = "RJafroc", mustWork = TRUE)
rdrArr1D <- DFReadSP(fileName)


}
