% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDataset.R
\name{isValidDataset}
\alias{isValidDataset}
\title{Check the validity of a dataset and FOM combination}
\usage{
isValidDataset(dataset, FOM, method = "OR", analysisOption = "RRRC")
}
\arguments{
\item{dataset}{The dataset object to be checked.}

\item{FOM}{The figure of merit.}

\item{method}{The analysis method, defaults to "OR".}

\item{analysisOption}{The desired generalization: "RRRC", "RRFC", "FRRC" or "ALL", 
defaults to "RRRC"}
}
\value{
\code{TRUE} if combination is valid, \code{FALSE} otherwise.
}
\description{
Checks the validity of a specified dataset and FOM combination.
}
