% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDataset.R
\name{isValidDataset}
\alias{isValidDataset}
\title{Check the validity of a dataset for FOM and other input parameters}
\usage{
isValidDataset(
  dataset,
  FOM,
  method = "OR",
  covEstMethod = "jackknife",
  analysisOption = "RRRC"
)
}
\arguments{
\item{dataset}{The dataset object to be checked.}

\item{FOM}{The figure of merit.}

\item{method}{The analysis method "OR" (default) or "DBM".}

\item{covEstMethod}{The covariance estimation method "jackknife" (default), 
"bootstrap" or "DeLong" (for an ROC dataset).}

\item{analysisOption}{Specification of the random factor(s): "RRRC" (default), "RRFC", 
or "FRRC.}
}
\value{
None.
}
\description{
Checks the validity of a specified dataset for FOM and other 
    input parameters.
}
