% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_LLF}
\alias{RSM_LLF}
\title{RSM predicted FROC ordinate}
\usage{
RSM_LLF(z, mu, nu)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the FROC ordinate.}

\item{mu}{The RSM mu parameter.}

\item{nu}{The RSM nu prime parameter.}
}
\value{
yFROC
}
\description{
RSM predicted FROC ordinate
}
\examples{
RSM_LLF(1,1,0.5)

}
