% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_NLF}
\alias{RSM_NLF}
\title{RSM predicted FROC abscissa}
\usage{
RSM_NLF(z, lambda)
}
\arguments{
\item{z}{The z-vector at which to evaluate the FROC abscissa.}

\item{lambda}{The scalar RSM lambda parameter.}
}
\value{
NLF, the abscissa of the FROC curve
}
\description{
RSM predicted FROC abscissa
}
\examples{
RSM_NLF(c(1,2),1)

}
