% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_pdfD}
\alias{RSM_pdfD}
\title{RSM predicted ROC-rating pdf for diseased cases}
\usage{
RSM_pdfD(z, mu, lambda, nu, lesDistr)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the pdf.}

\item{mu}{The mu parameter of the RSM.}

\item{lambda}{The (intrinsic) lambda parameter of the RSM.}

\item{nu}{The (intrinsic) nu parameter of the RSM.}

\item{lesDistr}{The lesion distribution 1D vector.}
}
\value{
pdf
}
\description{
RSM predicted ROC-rating pdf for diseased cases
}
\examples{
lesDistr <- c(0.5, 0.5)
RSM_pdfD(1,1,1,0.5, lesDistr)
lesDistr <- c(0.2, 0.3, 0.5)
RSM_pdfD(1,1,1,0.5, lesDistr)

}
