% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_pdfD}
\alias{RSM_pdfD}
\title{RSM predicted ROC-rating pdf for diseased cases}
\usage{
RSM_pdfD(z, mu, lambda, nu, lesDistr)
}
\arguments{
\item{z}{The z-vector at which to evaluate the pdf.}

\item{mu}{The scalar RSM mu parameter.}

\item{lambda}{The scalar RSM lambda parameter.}

\item{nu}{The scalar RSM nu parameter.}

\item{lesDistr}{The lesion distribution 1D vector.}
}
\value{
pdf, density function for diseased cases
}
\description{
RSM predicted ROC-rating pdf for diseased cases
}
\examples{
RSM_pdfD(c(1,2),1,1,0.9, c(0.5, 0.5))
RSM_pdfD(c(1,2),1,1,0.5, c(0.2, 0.3, 0.5))

}
