% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_pdfN}
\alias{RSM_pdfN}
\title{RSM predicted ROC-rating pdf for non-diseased cases}
\usage{
RSM_pdfN(z, lambda)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the pdf.}

\item{lambda}{The (physical) lambda parameter of the RSM.}
}
\value{
pdf
}
\description{
RSM predicted ROC-rating pdf for non-diseased cases
}
\examples{
RSM_pdfN(1,1)

}
