% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_wLLF}
\alias{RSM_wLLF}
\title{RSM predicted wAFROC ordinate}
\usage{
RSM_wLLF(zeta, mu, nu, lesDistr, relWeights)
}
\arguments{
\item{zeta}{The zeta value at which to evaluate the FROC ordinate.}

\item{mu}{The RSM mu parameter.}

\item{nu}{The RSM nu prime parameter.}

\item{lesDistr}{Lesion distribution vector.}

\item{relWeights}{Relative lesion weights vector.}
}
\value{
wLLF
}
\description{
RSM predicted wAFROC ordinate
}
\examples{
RSM_wLLF(1,1,0.5, lesDistr = c(0.5, 0.4, 0.1), relWeights = c(0.7, 0.2, 0.1)) 


}
