% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmFormulae.R
\name{RSM_xFROC}
\alias{RSM_xFROC}
\title{RSM predicted FROC abscissa}
\usage{
RSM_xFROC(z, lambdaP)
}
\arguments{
\item{z}{The z-sample value at which to evaluate the FROC abscissa.}

\item{lambdaP}{The (physical) lambda prime parameter of the RSM.}
}
\value{
xFROC
}
\description{
RSM predicted FROC abscissa
}
\examples{
RSM_pdfN(1,1)

}
