% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateFrocDataset.R
\name{SimulateFrocDataset}
\alias{SimulateFrocDataset}
\title{Simulates an MRMC uncorrelated FROC dataset using the RSM}
\usage{
SimulateFrocDataset(
  mu,
  lambda,
  nu,
  zeta1,
  I,
  J,
  K1,
  K2,
  perCase,
  seed = NULL,
  deltaMu = 0
)
}
\arguments{
\item{mu}{mu parameter of the RSM}

\item{lambda}{RSM lambda parameter}

\item{nu}{RSM nu parameter}

\item{zeta1}{Lowest reporting threshold}

\item{I}{Number of treatments, default is 1}

\item{J}{Number of readers}

\item{K1}{Number of non-diseased cases}

\item{K2}{Number of diseased cases}

\item{perCase}{A K2 length array containing the numbers of lesions per diseased case}

\item{seed}{Initial seed for random number generator, default 
\code{NULL}, for random seed.}

\item{deltaMu}{Inter-treatment increment in mu, default zero}
}
\value{
An FROC dataset.
}
\description{
Simulates an uncorrelated MRMC FROC dataset for specified numbers of
   readers and treatments
}
\details{
See book chapters on the Radiological Search Model (RSM) for details. 
   In this code correlations between ratings on the same case are assumed to be zero.
}
\examples{
set.seed(1) 
K1 <- 5;K2 <- 7;
maxLL <- 2;perCase <- floor(runif(K2, 1, maxLL + 1))
mu <- 1;lambda <- 1;nu <- 0.99 ;zeta1 <- -1
I <- 2; J <- 5

frocDataRaw <- SimulateFrocDataset(
  mu = mu, lambda = lambda, nu = nu, zeta1 = zeta1,
  I = I, J = J, K1 = K1, K2 = K2, perCase = perCase )
  
## plot the data
ret <- PlotEmpiricalOperatingCharacteristics(frocDataRaw, opChType = "FROC")
## print(ret$Plot)

}
\references{
Chakraborty DP (2017) \emph{Observer Performance Methods for Diagnostic Imaging - Foundations, 
Modeling, and Applications with R-Based Examples}, CRC Press, Boca Raton, FL. 
\url{https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840}
}
