% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSignificanceTestingCrossedModalities.R
\name{StSignificanceTestingCrossedModalities}
\alias{StSignificanceTestingCrossedModalities}
\title{Perform significance testing using crossed treatments analysis}
\usage{
StSignificanceTestingCrossedModalities(
  crossedData,
  avgIndx,
  FOM = "wAFROC",
  alpha = 0.05,
  option = "ALL"
)
}
\arguments{
\item{crossedData}{The crossed treatments dataset}

\item{avgIndx}{The index of the treatment to be averaged over}

\item{FOM}{See \code{\link{StSignificanceTesting}}}

\item{alpha}{See \code{\link{StSignificanceTesting}}}

\item{option}{See \code{\link{StSignificanceTesting}}}
}
\value{
The return list contains the same items with \code{\link{StSignificanceTesting}}.
}
\description{
Performs ORH analysis for specified crossed treatments dataset 
   averaged over specified treatment factor
}
\examples{
\donttest{ 
## read the raw data file in extdata directory
crossedFileName <- system.file("extdata", "CrossedModalitiesData.xlsx", 
package = "RJafroc", mustWork = TRUE)
crossedData <- DfReadCrossedModalities(crossedFileName)
retCrossed1 <- StSignificanceTestingCrossedModalities(crossedData, 1)

## read the built in dataset
retCrossed2 <- StSignificanceTestingCrossedModalities(datasetCrossedModality, 1)
}

}
