% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSP.R
\name{StSP}
\alias{StSP}
\title{Performs OR significance testing for SPLIT-PLOT A or C datasets}
\usage{
StSP(dataset, FOM, alpha = 0.05, analysisOption = "RRRC")
}
\arguments{
\item{dataset}{The dataset to be analyzed, see \code{\link{RJafroc-package}}. 
Must have two or more treatments and two or more readers. The dataset design
must be "SPLIT-PLOT-A" or "SPLIT-PLOT-C".}

\item{FOM}{The figure of merit}

\item{alpha}{The significance level of the test, default is 0.05}

\item{analysisOption}{Determines which factors are regarded as random vs. fixed:
\itemize{ 
   \item \code{"RRRC"} = random-reader random case, the default,
   \item \code{"FRRC"} = fixed-reader random case, 
   \item \code{"RRFC"} = random-reader fixed case, 
}}
}
\value{
Results of the analysis
}
\description{
Performs Obuchowski-Rockette (OR) significance testing for 
   specified dataset.
}
\examples{
fileName <- system.file("extdata", "/toyFiles/ROC/rocSpAZP.xlsx", 
package = "RJafroc", mustWork = TRUE)
dsSpA <- DfReadSP(fileName)
ret <- StSP(dsSpA, FOM = "Wilcoxon")

ret <- StSP(datasetFROCSpC, FOM = "wAFROC")


}
