% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilCnvtDbmOrVarComp.R
\name{UtilDBM2ORVarComp}
\alias{UtilDBM2ORVarComp}
\title{Convert between DBM and OR variance components}
\usage{
UtilDBM2ORVarComp(K, DBMVarComp)
}
\arguments{
\item{\code{K}}{Total number of cases}

\item{\code{DBMVarComp}}{DBM variance components, a list or data.frame 
containing varR, varC, varTR, varTC, varRC and varErr}

\item{\code{ORVarComp}}{OR variance components, a list or data.frame 
containing varR, varTR, Cov1, Cov2, Cov3 and Var}
}
\value{
\code{UtilDBM2ORVarComp} returns the equivalent OR variance components

\code{UtilOR2DBMVarComp} returns the equivalent DBM variance components
}
\description{
\code{UtilDBM2ORVarComp} converts from DBM variance components to OR 
   variance components, while \code{UtilOR2DBMVarComp} converts from
   OR to DBM variance components.
}
\examples{
DBMVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "DBMH")$ANOVA$VarCom
UtilDBM2ORVarComp(114, DBMVarCom)

ORVarCom <- StSignificanceTesting(dataset02, FOM = "Wilcoxon", method = "ORH")$ANOVA$VarCom
UtilOR2DBMVarComp(114, ORVarCom)


}
