% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilLesionDistr.R
\name{UtilLesionDistr}
\alias{UtilLesionDistr}
\title{Lesion distribution in the dataset}
\usage{
UtilLesionDistr(dataset)
}
\arguments{
\item{dataset}{The dataset}
}
\value{
The lesion distribution array
}
\description{
The lesion distribution for a dataset.
}
\details{
Two characteristics of an FROC dataset, apart from the 
   ratings, affect the FOM: the distribution of lesion per case and the 
   distribution of lesion weights. This function addresses the first. 
   The weights are addressed by \link{UtilLesionWeightsDistr}. \code{lesDistr} 
   is an [1:nRow,2] array, where \code{nRow} is the number of \bold{unique} 
   values of lesions per case in the dataset. The first column of the array contains 
   the number of lesions per case. The second column contains the corresponding 
   fraction of diseased cases. See \link{PlotRsmOperatingCharacteristics} for a 
   function that depends on \code{lesDistr}. See Chapter00Vignette2 for more details.
}
\examples{
UtilLesionDistr (dataset01) # FROC data
UtilLesionDistr (dataset02) # ROC data
UtilLesionDistr (datasetCadLroc) # LROC data

}
