% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilLesionDistr.R
\name{UtilLesionDistr}
\alias{UtilLesionDistr}
\title{Lesion distribution  matrix}
\usage{
UtilLesionDistr(dataset)
}
\arguments{
\item{dataset}{The supplied dataset}
}
\value{
The lesion distribution matrix
}
\description{
Extracts the lesion distribution matrix for a dataset.
}
\details{
The lesion distribution matrix has Lmax rows and two columns. 
   The first column contains the integers 1, 2, ..., Lmax and the second 
   column contains the fraction of diseased cases with the number of lesions 
   per case specified in the first column.
}
\examples{
UtilLesionDistr (dataset01) # FROC data
UtilLesionDistr (dataset02) # ROC data
UtilLesionDistr (datasetCadLroc) # LROC data

}
