% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilLesionWeightsDistr.R
\name{UtilLesionWeightsDistr}
\alias{UtilLesionWeightsDistr}
\title{Lesion weights distribution}
\usage{
UtilLesionWeightsDistr(dataset)
}
\arguments{
\item{dataset}{The supplied dataset.}
}
\value{
The lesion weights distribution, see \code{\link{PlotRsmOperatingCharacteristics}}
}
\description{
Computes the lesion weights  distribution, assuming equal weights.
}
\examples{
UtilLesionWeightsDistr (dataset01) # FROC data
UtilLesionWeightsDistr (dataset02) # ROC data
UtilLesionWeightsDistr (datasetCadLroc) # LROC data

}
