% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilOrVarCov.R
\name{UtilOrVarCov}
\alias{UtilOrVarCov}
\title{Obuchowski-Rockette variance components for dataset}
\usage{
UtilOrVarCov(
  dataset,
  FOM,
  FPFValue = 0.2,
  covEstMethod = "jackknife",
  nBoots = 200,
  seed = NULL
)
}
\arguments{
\item{dataset}{Factorial or cross-modality dataset}

\item{FOM}{Figure of merit}

\item{FPFValue}{Only needed for \code{LROC} data \strong{and} FOM = "PCL" or "ALROC":
the \code{FPFValue} at which to evaluate a partial curve based figure of merit. 
The default is \code{FPFValue} = 0.2.}

\item{covEstMethod}{The covariance estimation method, "jackknife" 
(the default) or "bootstrap" or "DeLong" ("DeLong" is applicable only for 
FOM = "Wilcoxon").}

\item{nBoots}{Only needed for bootstrap covariance estimation method. The number 
of bootstraps, defaults to 200.}

\item{seed}{Only needed for the bootstrap covariance estimation method. The initial 
seed for the random number generator, the default is \code{NULL}, for random seed.}
}
\value{
A list containing the following \code{data.frames}: 
    \itemize{
    \item{\code{foms}}: the figures of merit for different modality-reader combinations 
    \item{\code{TRanova}}: the OR modality-reader ANOVA table 
    \item{\code{VarCom}}: the OR variance-components \code{Cov1}, \code{Cov2}, 
    \code{Cov3}, \code{Var} and correlations \code{rho1}, \code{rho2} and \code{rho3} 
    \item{\code{IndividualTrt}}: the individual modality mean-squares, \code{Var} and \code{Cov2} values
    \item{\code{IndividualRdr}}: the individual reader mean-squares, \code{Var} and \code{Cov1} values
    }
}
\description{
Obuchowski-Rockette variance components for dataset
}
\details{
The variance components are identical to those obtained using 
    \link{StSignificanceTesting} with \code{method = "OR"}.
}
\examples{
## use the default jackknife for covEstMethod
vc <- UtilOrVarCov(dataset02, FOM = "Wilcoxon")
str(vc) 

UtilOrVarCov(dataset02, FOM = "Wilcoxon", 
   covEstMethod = "bootstrap", nBoots = 2000, seed = 100)$VarCom 

UtilOrVarCov(dataset02, FOM = "Wilcoxon", covEstMethod = "DeLong")$VarCom

UtilOrVarCov(datasetXModality, FOM = "wAFROC") 
  
}
