% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilOutputReport.R
\name{UtilOutputReport}
\alias{UtilOutputReport}
\title{Generate a text- or Excel-formatted report file (one-treatment factorial datasets only)}
\usage{
UtilOutputReport(
  dataset,
  FOM,
  method = "OR",
  covEstMethod = "jackknife",
  analysisOption = "ALL",
  ReportFileExt = "txt",
  ReportFileBaseName = NULL,
  alpha = 0.05,
  nBoots = 200,
  sequentialNames = FALSE,
  overWrite = FALSE
)
}
\arguments{
\item{dataset}{The dataset object to be analyzed, \emph{not the Excel input 
file name!}, see \code{Dataset} in \code{\link{RJafroc-package}}. 
This \code{R}-object is created by \code{\link{DfReadDataFile}}.}

\item{FOM}{The figure of merit; see \code{\link{St}}.}

\item{method}{The significance testing method, \code{"DBM"} or \code{"OR"} 
(the default).}

\item{covEstMethod}{The default is "Jackknife"; only needed 
for method = \code{"OR"} when "Bootstrap" or "DeLong" are possible.}

\item{analysisOption}{"RRRC", "FRRC", "RRFC or "ALL" (the default); 
see \code{\link{St}}.}

\item{ReportFileExt}{The file extension determines the type of output. 
\code{"txt"} (default) for a text file or \code{"xlsx"} 
for an Excel file.}

\item{ReportFileBaseName}{The report file \bold{base name}, i.e., without the extension (which is 
specified by option \code{ReportFileExt})). The output file is created 
\bold{in the user's directory}. The default is \code{NULL}, 
see \code{details}.}

\item{alpha}{See \code{\link{St}}; the default is 0.05.}

\item{nBoots}{See \code{\link{St}}; only needed for \code{"OR"} analysis; 
the default is 200.}

\item{sequentialNames}{A logical variable: if \code{TRUE}, consecutive integers 
(starting from 1) will be used as the modality and reader IDs in the 
output report. Otherwise, modality and reader IDs in the original dataset 
are used. This option is needed for aesthetics, as long names can mess
up the output. The default is \code{FALSE.}}

\item{overWrite}{A \code{logical} variable: if \code{FALSE}, a warning will 
be issued if the report file already exists and the program will wait 
until the user inputs "y" or "n" to determine whether to overwrite the 
existing file. If \code{TRUE}, the existing file will be silently overwritten. 
The default is \code{FALSE}.}
}
\value{
\code{StResult}: the object returned by \code{\link{St}}.
}
\description{
Generates a formatted report resulting from the analysis, for a
   factorial dataset, and saves it to a text- or an Excel-formatted file.
}
\details{
This function is available only for factorial datasets. 
\code{ReportFileBaseName}: The report file \bold{base name}. The default value
    \code{NULL} generates a temporary text file, whose very long name is 
    displayed. However, the file is very hard to locate. This is so that the 
    package passes CRAN checks, as writing to the project directory, or any 
    of its sub-directories, is frowned upon.
}
\examples{

\donttest{
 # text output is created in a temporary file
ret <- UtilOutputReport(dataset03, FOM = "Wilcoxon")
 # Excel output is created in a temporary file
ret <- UtilOutputReport(dataset03, FOM = "Wilcoxon", ReportFileExt = "xlsx") 

}
       
}
