% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilOutputReport.R
\name{UtilOutputReport}
\alias{UtilOutputReport}
\title{Generate a formatted report file}
\usage{
UtilOutputReport (dataset, DataFileName, DataFileFormat = "JAFROC", delimiter = ",", dataset, 
   dataDescription = deparse(substitute(dataset)), ReportFileName, ReportFileFormat = "txt",
   method = "DBMH", FOM = "wJAFROC", alpha = 0.05, covEstMethod = "Jackknife", 
   nBoots = 200, renumber = FALSE, overwrite = TRUE)
}
\arguments{
\item{dataset}{The dataset object to be analyzed (not the file name), \link{RJafroc-package}.}

\item{DataFileName}{A string variable containing the name of the data file 
to be analyzed, see \link{DfReadDataFile} and "Details".}

\item{DataFileFormat}{The format of the data specified in \code{DataFileName}: 
see \link{DfReadDataFile} and "Details".}

\item{delimiter}{See \link{DfReadDataFile}.}

\item{dataDescription}{Only needed if a dataset is not specified. 
It is a string descriptor of the dataset; the default is the variable 
name of \code{dataset}}

\item{ReportFileName}{The file name of the output report file. If this parameter 
is missing, the function will use \code{DataFileName} or \code{dataDescription} 
followed by the underscore separated concatenation of \code{method} 
and \code{FOM} as the output report file.}

\item{ReportFileFormat}{The format of the output report. The two available formats are 
"txt" and "xlsx", which correspond to a formatted text file and an 
Excel file, respectively. "txt" is the default.}

\item{method}{The analysis method: \code{"ORH"} or \code{"DBMH"}.}

\item{FOM}{See \link{StSignificanceTesting}.}

\item{alpha}{See \link{StSignificanceTesting}.}

\item{covEstMethod}{See \link{StSignificanceTesting}.}

\item{nBoots}{See \link{StSignificanceTesting}.}

\item{renumber}{A logical variable: if \code{TRUE}, consecutive integers 
(staring from 1) will be used as the modality and reader IDs in the 
output report. Otherwise, modality and reader IDs in the original data 
file will be used. This option may be needed for aesthetics.}

\item{overwrite}{A \code{logical} variable: if \code{FALSE}, a warning will 
be issued if the report file already exists and the program will wait 
until the user inputs "y" or "n" to determine whether to overwrite the 
existing file. If \code{TRUE}, an existing file will be silently overwritten.}
}
\value{
A formatted report of the data analysis, patterned roughly on that of 
   OR-DBM MRMC V2.5.
}
\description{
Generate a formatted report of the analysis and save to a text file
}
\details{
At least one of the combinations of \code{DataFileName} and \code{DataFileFormat} 
   or \code{dataset} and \code{dataDescription} must be specified. 
   If both are specified, the data file \code{DataFileName} is analyzed and 
   the \code{dataset} is ignored.
}
\examples{
\dontrun{
UtilOutputReport(dataset = dataset02, method = "DBMH", FOM = "Wilcoxon", 
             dataDescription = "MyROCData", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, method = "DBMH", FOM = "Wilcoxon", 
dataDescription = "MyROCData",ReportFileFormat = "xlsx", overwrite = TRUE)

## Generate a analysis report for a data file.
fn <- system.file("extdata", "includedRocData.xlsx", 
package = "RJafroc", mustWork = TRUE)
UtilOutputReport(DataFileName = fn, method = "DBMH", FOM = "Wilcoxon",
             overwrite = TRUE)
             
## Output report for an existing dataset
UtilOutputReport(dataset = dataset02, method = "DBMH", FOM = "Wilcoxon", 
             ReportFileName = "MyROCDataAnalysis.txt") 
UtilOutputReport(dataset = dataset02, method = "ORH", FOM = "Wilcoxon", overwrite = TRUE)
## UtilOutputReport(dataset = dataset05, method = "DBMH", FOM = "Wilcoxon") # ERROR!
UtilOutputReport(dataset = dataset05, method = "ORH") # default FOM is wJAFROC
UtilOutputReport(dataset = dataset05, method = "DBMH", FOM = "HrAuc")
}
       
}
