% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilOutputReport.R
\name{UtilOutputReport}
\alias{UtilOutputReport}
\title{Generate a formatted report file}
\usage{
UtilOutputReport(dataset, DataFileName, DataFileFormat, delimiter = ",",
  dataDescription = "MyData", ReportFileName, ReportFileFormat = "txt",
  method = "DBMH", FOM = "Wilcoxon", alpha = 0.05,
  covEstMethod = "Jackknife", nBoots = 200, renumber = FALSE,
  overwrite = TRUE)
}
\arguments{
\item{dataset}{The dataset object to be analyzed (not the file name), see \code{Dataset} 
in \code{\link{RJafroc-package}}.}

\item{DataFileName}{A string variable containing the name of the data file 
to be analyzed, in the form \code{baseName.ext}, see \code{\link{DfReadDataFile}}.}

\item{DataFileFormat}{The format of the data specified in \code{DataFileName}: 
see \code{\link{DfReadDataFile}}. Must be specified if \code{DataFileName}
is specified. Valid values are "JAFROC", "MRMC" or "iMRMC"}

\item{delimiter}{used in text input data file, See \code{\link{DfReadDataFile}}.}

\item{dataDescription}{Optional description of the data, default is \code{"MyData"}}

\item{ReportFileName}{The file name of the output report file. If this parameter 
is missing, the function will use \code{DataFileName} or \code{dataDescription} 
followed by the underscore separated concatenation of \code{method} 
and \code{FOM} as the output report file.}

\item{ReportFileFormat}{The format of the output report. The two available formats are 
\code{"txt"} (the default) and \code{"xlsx"}, corresponding to a formatted text file or an 
Excel file, respectively.}

\item{method}{The significance testing method, \code{"ORH"} or (the default) \code{"DBMH"}.}

\item{FOM}{The figure of merit; see \code{\link{StSignificanceTesting}}.}

\item{alpha}{See \code{\link{StSignificanceTesting}}.}

\item{covEstMethod}{See \code{\link{StSignificanceTesting}}; only needed for \code{"ORH"} analysis.}

\item{nBoots}{See \code{\link{StSignificanceTesting}}; only needed for \code{"ORH"} analysis.}

\item{renumber}{A logical variable: if \code{TRUE}, consecutive integers 
(staring from 1) will be used as the treatment and reader IDs in the 
output report. Otherwise, treatment and reader IDs in the original data 
file will be used. This option may be needed for aesthetics.}

\item{overwrite}{A \code{logical} variable: if \code{FALSE}, a warning will 
be issued if the report file already exists and the program will wait 
until the user inputs "y" or "n" to determine whether to overwrite the 
existing file. If \code{TRUE}, an existing file will be silently overwritten.}
}
\value{
sigTestResult The object returned by \code{\link{StSignificanceTesting}}.
}
\description{
Generate a formatted report of the analysis and save to a text file
}
\details{
A formatted report of the data analysis, patterned roughly on that of 
   OR-DBM MRMC V2.5, is written to the output file.
}
\examples{

\dontrun{
# moved here since it was writing MyDataOutput.txt to 
project root directory, which was caught by Ubuntu (#13) and CRAN
UtilOutputReport(dataset = dataset03, overwrite = TRUE)

## Generate reports for a dataset object
UtilOutputReport(dataset = dataset02, method = "DBMH", 
             dataDescription = "MyROCData1", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, method = "DBMH", 
dataDescription = "MyROCData2",ReportFileFormat = "xlsx", overwrite = TRUE)

UtilOutputReport(dataset = dataset02, method = "ORH", 
             dataDescription = "MyROCData3", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, method = "ORH", 
dataDescription = "MyROCData4",ReportFileFormat = "xlsx", overwrite = TRUE)

## Generate report for a data file
fn <- system.file("extdata", "includedRocData.xlsx", 
package = "RJafroc", mustWork = TRUE)
UtilOutputReport(DataFileName = fn, DataFileFormat = "JAFROC", method = "DBMH",
             overwrite = TRUE, ReportFileFormat = "xlsx")
             
## Output report for an existing dataset
## UtilOutputReport(dataset05, method = "DBMH") # ERROR as FOM is 
   incompatible with FROC data

UtilOutputReport(dataset05, method = "ORH", FOM = "wAFROC")

UtilOutputReport(dataset05, method = "DBMH", FOM = "HrAuc")

UtilOutputReport(dataset05, method = "DBMH", FOM = "HrAuc", ReportFileFormat = "xlsx")
}
       
}
