% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilSpecifyLesionWeightsDistr.R
\name{UtilSpecifyLesionWeightsDistr}
\alias{UtilSpecifyLesionWeightsDistr}
\title{Specify lesion weight distribution array}
\usage{
UtilSpecifyLesionWeightsDistr(specified1DLesDistr, relWeights = 0)
}
\arguments{
\item{specified1DLesDistr}{The specified lesion distribution array. For example, 
c(0.1, 0.2, 0, 0.7) specifies 10 percent of diseased cases have one lesion,
20 percent have two lesions, none have 3 lesions and 70 percent have 
four lesions. See example below.}

\item{relWeights}{The relative weights of the lesions; a vector of 
length maxLL; if zero, the default, equal weights are assumed.}
}
\value{
lesWghtDistr The lesion weights distribution array
}
\description{
{Specify the lesion weights, the underlying assumption being that 
   lesion 1 is the same type across all diseased cases, lesion 2 is the same 
   type across all diseased cases, ..., etc. This allows assignment of weights 
   independent of the case index.}
}
\note{
This function is distinct from `UtilLesionWeightsDistr()` which gets 
   the lesion weights array for a dataset.
}
\examples{
specified1DLesDistr <- c(0.1, 0.2, 0, 0.7)
relWeights <- c(0.2, 0.4, 0.1, 0.3)
UtilSpecifyLesionWeightsDistr (specified1DLesDistr, relWeights) 



}
