% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilVarComponents.R
\name{UtilVarComponentsOR}
\alias{UtilVarComponentsOR}
\title{Utility for Obuchowski-Rockette variance components}
\usage{
UtilVarComponentsOR(
  dataset,
  FOM,
  FPFValue = 0.2,
  covEstMethod = "jackknife",
  nBoots = 200,
  seed = NULL
)
}
\arguments{
\item{dataset}{The dataset object}

\item{FOM}{The figure of merit}

\item{FPFValue}{Only needed for \code{LROC} data \strong{and} FOM = "PCL" or "ALROC";
where to evaluate a partial curve based figure of merit. The default is 0.2.}

\item{covEstMethod}{The covariance estimation method, "jackknife" 
(the default) or "bootstrap".}

\item{nBoots}{Only needed for bootstrap covariance estimation method. The number 
of bootstraps, defaults to 200.}

\item{seed}{Only needed for the bootstrap covariance estimation method. The initial 
seed for the random number generator, the default is \code{NULL}, as if no seed 
has been specified.}
}
\value{
A list object containing the variance components.
}
\description{
Utility for Obuchowski-Rockette variance components
}
\details{
The variance components are obtained using \link{StSignificanceTesting} 
    with \code{method = "ORH"}.
}
\examples{
UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$varComp 
## uses the default jackknife for covEstMethod

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon", 
   covEstMethod = "bootstrap", nBoots = 2000, seed = 100)$varComp 

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon", 
   covEstMethod = "DeLong")$varComp 
  
}
